# backend/api/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/users": {
            "get": {
                "operationId": "UsersController_findAll",
                "parameters": [
                    {
                        "name": "from",
                        "required": false,
                        "in": "query",
                        "description": "Filter by this date",
                        "schema": {
                            "format": "2021-01-01T11:59:59.999Z",
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "required": false,
                        "in": "query",
                        "description": "Filter by this date",
                        "schema": {
                            "format": "2021-01-01T11:59:59.999Z",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserEntity"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "operationId": "UsersController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "delete": {
                "operationId": "UsersController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
        "/questions/multiple-choice": {
            "post": {
                "operationId": "QuestionsController_createMultipleChoice",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMultipleChoiceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "questions"
                ]
            }
        },
        "/questions/true-false": {
            "post": {
                "operationId": "QuestionsController_createTrueFalse",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTrueFalseDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "questions"
                ]
            }
        },
        "/questions/short-answer": {
            "post": {
                "operationId": "QuestionsController_createShortAnswer",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateShortAnswerDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "questions"
                ]
            }
        },
        "/questions/multiple-select": {
            "post": {
                "operationId": "QuestionsController_createMultipleSelect",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMultipleChoiceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "questions"
                ]
            }
        },
        "/questions/{id}": {
            "get": {
                "operationId": "QuestionsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "questions"
                ]
            },
            "patch": {
                "operationId": "QuestionsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateQuestionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "questions"
                ]
            },
            "delete": {
                "operationId": "QuestionsController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "questions"
                ]
            }
        }
    },
    "info": {
        "title": "NestJS API 0.0.1}",
        "description": "The NestJS API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "http",
                "in": "header",
                "name": "auth-cookie",
                "description": "Cookies are set automatically. No need to add.",
                "scheme": "bearer"
            },
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://localhost:3000/auth/google/backend-login",
                        "scopes": {
                            "read:users": "read users",
                            "write:users": "modify users"
                        }
                    }
                }
            }
        },
        "schemas": {
            "UserEntity": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "USER"
                        ],
                        "default": "USER"
                    },
                    "id": {
                        "type": "number"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "emailVerified": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "role",
                    "id",
                    "email",
                    "firstName",
                    "lastName",
                    "image",
                    "emailVerified"
                ]
            },
...
            "CreateMultipleChoiceDto": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "tags": {
                        "default": [],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "question": {
                        "type": "string"
                    },
                    "answer": {
                        "type": "number"
                    },
                    "examId": {
                        "type": "number"
                    }
                },
                "required": [
                    "options",
                    "tags",
                    "question",
                    "answer",
                    "examId"
                ]
            },
            "CreateTrueFalseDto": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "tags": {
                        "default": [],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "question": {
                        "type": "string"
                    },
                    "answer": {
                        "type": "boolean"
                    },
                    "examId": {
                        "type": "number"
                    }
                },
                "required": [
                    "tags",
                    "question",
                    "answer",
                    "examId"
                ]
            },
            "CreateShortAnswerDto": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "tags": {
                        "default": [],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "question": {
                        "type": "string"
                    },
                    "answer": {
                        "type": "string"
                    },
                    "examId": {
                        "type": "number"
                    }
                },
                "required": [
                    "tags",
                    "question",
                    "answer",
                    "examId"
                ]
            },
            "UpdateQuestionDto": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "tags": {
                        "default": [],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "question": {
                        "type": "string"
                    },
                    "type": {
                        "type": "object"
                    },
                    "examId": {
                        "type": "number"
                    }
                }
            }
        }
    }
}